move document highlighter to be part of registry
